refactor : 변경된 폴더 구조 적용
